Simplify layout by removing environment-specific CSS and JS
The commit removes the distinction between development and other environments for CSS and Javascript in the _Layout.cshtml file. It simplifies the page layout by consistently using local versions of Bootstrap and Font-Awesome, and by modifying the placement of certain elements. All environments will now use the same CSS and JS, which streamlines testing and deployment.