Upgrading an application from ASP.NET Core 2.2 to 8.0 (#8)
* Update UI elements styles in Index.cshtml

In the PhoneBook Index view, changes have been made in the style of the badge and buttons to partially fix their appearance. This included modifying the background color for the badge and varying the width and icon dimensions of the Edit, Details, and Delete buttons.

* Simplify layout by removing environment-specific CSS and JS

The commit removes the distinction between development and other environments for CSS and Javascript in the _Layout.cshtml file. It simplifies the page layout by consistently using local versions of Bootstrap and Font-Awesome, and by modifying the placement of certain elements. All environments will now use the same CSS and JS, which streamlines testing and deployment.

* Add check for non-empty model in Index view

This commit includes a check to ensure that the list provided to the Index view of the PhoneBook is not empty. Previously, if the list was empty, the view would attempt to take the first entry and fail. The if statement ensures this error is avoided.

* Refactor search functionality in PhoneBookController

Search string validation was simplified and applied directly to the database query rather than to an in-memory list of entries. Further, the ordering and pagination of entries were also shifted to the query level, improving efficiency and performance.

* Refactor database connection setup and enhance options configuration

Removed excessive use of `MySqlConnector` and refactored the way connection strings are retrieved. Also applied cleaner structuring for adding database context services. Added configuration for `DefaultUserOptions` and conditional addition of `DatabaseDeveloperPageExceptionFilter` based on the environment.